# api/openapi.v2.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v2/transactions": {
      "get": {
        "operationId": "Transactions_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based; 1 by default.",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "50 by default, at most 200.",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "type": "number"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "categoryIds",
            "required": false,
            "in": "query",
            "description": "Several ids, comma separated: `?categoryIds=3,7`. Each brings its whole branch.",
            "schema": {
              "pattern": "^\\d+(,\\d+)*$",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["expense", "income", "transfer"]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["cleared", "pending"]
            }
          },
          {
            "name": "tagId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text search over description, merchant and notes.",
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "name": "minAmount",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxAmount",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "`date`, `amount`, `createdAt` or `merchant`; `-` in front for descending.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "meta"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transaction"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/TransactionPageMeta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The input is invalid; `errors` names each field at fault. Codes: `bad_request`, `invalid_fields`, `invalid_id`, `user_already_active`, `cannot_target_self`, `last_active_admin`, `credit_fields_on_non_credit`, `no_files`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token (or session to renew). Codes: `unauthenticated`, `invalid_credentials`, `invalid_token`, `session_expired`, `session_revoked`, `wrong_current_password`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached; retry later. Codes: `rate_limited`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. Codes: `internal_error`, `identity_provider_failed`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["TransactionsV2"]
      },
      "post": {
        "operationId": "Transactions_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "meta"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Transaction"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {},
                      "maxProperties": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The input is invalid; `errors` names each field at fault. Codes: `bad_request`, `invalid_fields`, `invalid_id`, `user_already_active`, `cannot_target_self`, `last_active_admin`, `credit_fields_on_non_credit`, `no_files`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token (or session to renew). Codes: `unauthenticated`, `invalid_credentials`, `invalid_token`, `session_expired`, `session_revoked`, `wrong_current_password`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not owned by this user. Codes: `not_found`, `receipt_file_missing`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (a duplicate, a resource in use). Codes: `conflict`, `duplicate`, `constraint_violation`, `account_has_transactions`, `category_has_children`, `reassignment_required`, `reassignment_target_inside`, `template_requires_empty`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Well-formed but breaks a business rule. Codes: `validation_failed`, `check_violation`, `weak_password`, `password_reset_not_managed`, `splits_unbalanced`, `amount_breaks_splits`, `transfer_same_account`, `transfer_leg_locked`, `account_not_owned`, `category_not_owned`, `category_cycle`, `category_too_deep`, `multi_payment_requires_concept`, `multi_payment_requires_recurring`, `multi_payment_excludes_auto_paid`, `merge_into_itself`, `merge_requires_concepts`, `merge_source_has_children`, `interpretation_needs_text`, `concept_archived`, `cost_center_cannot_classify`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached; retry later. Codes: `rate_limited`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. Codes: `internal_error`, `identity_provider_failed`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["TransactionsV2"]
      }
    },
...
    "/api/v2/transactions/interpret": {
      "post": {
        "operationId": "Interpretacion_interpret",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "meta"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Interpretation"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {},
                      "maxProperties": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The input is invalid; `errors` names each field at fault. Codes: `bad_request`, `invalid_fields`, `invalid_id`, `user_already_active`, `cannot_target_self`, `last_active_admin`, `credit_fields_on_non_credit`, `no_files`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token (or session to renew). Codes: `unauthenticated`, `invalid_credentials`, `invalid_token`, `session_expired`, `session_revoked`, `wrong_current_password`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Well-formed but breaks a business rule. Codes: `validation_failed`, `check_violation`, `weak_password`, `password_reset_not_managed`, `splits_unbalanced`, `amount_breaks_splits`, `transfer_same_account`, `transfer_leg_locked`, `account_not_owned`, `category_not_owned`, `category_cycle`, `category_too_deep`, `multi_payment_requires_concept`, `multi_payment_requires_recurring`, `multi_payment_excludes_auto_paid`, `merge_into_itself`, `merge_requires_concepts`, `merge_source_has_children`, `interpretation_needs_text`, `concept_archived`, `cost_center_cannot_classify`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached; retry later. Codes: `rate_limited`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. Codes: `internal_error`, `identity_provider_failed`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["InterpretacionV2"]
      }
    },
    "/api/v2/transactions/capture": {
      "post": {
        "operationId": "Interpretacion_capture",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "meta"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Capture"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {},
                      "maxProperties": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The input is invalid; `errors` names each field at fault. Codes: `bad_request`, `invalid_fields`, `invalid_id`, `user_already_active`, `cannot_target_self`, `last_active_admin`, `credit_fields_on_non_credit`, `no_files`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token (or session to renew). Codes: `unauthenticated`, `invalid_credentials`, `invalid_token`, `session_expired`, `session_revoked`, `wrong_current_password`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not owned by this user. Codes: `not_found`, `receipt_file_missing`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (a duplicate, a resource in use). Codes: `conflict`, `duplicate`, `constraint_violation`, `account_has_transactions`, `category_has_children`, `reassignment_required`, `reassignment_target_inside`, `template_requires_empty`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Well-formed but breaks a business rule. Codes: `validation_failed`, `check_violation`, `weak_password`, `password_reset_not_managed`, `splits_unbalanced`, `amount_breaks_splits`, `transfer_same_account`, `transfer_leg_locked`, `account_not_owned`, `category_not_owned`, `category_cycle`, `category_too_deep`, `multi_payment_requires_concept`, `multi_payment_requires_recurring`, `multi_payment_excludes_auto_paid`, `merge_into_itself`, `merge_requires_concepts`, `merge_source_has_children`, `interpretation_needs_text`, `concept_archived`, `cost_center_cannot_classify`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached; retry later. Codes: `rate_limited`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. Codes: `internal_error`, `identity_provider_failed`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["InterpretacionV2"]
      }
    },
...
  },
  "info": {
    "title": "Coco API",
    "description": "Contract v2, served under `/api/v2`: the v1 API in English and camelCase, with every\nlist paginated as `{ data, meta: { page, perPage, total } }`.\n\nEvery successful JSON response is wrapped as `{ data, meta }`. Every error is\n`application/problem+json` (RFC 9457): `{ type, title, status, detail, code, errors? }`.\nSwitch on `code`: stable, in English, one per business rule. `errors` lists the fields\nthat failed.\nMessages meant for the user are in Spanish.\n\nIdentifiers are integers (int64) and money travels as a decimal string.",
    "version": "2",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "Identifies the kind of problem: `https://dev-cocoapp.viteri.me/problems/<code>`.",
            "example": "https://dev-cocoapp.viteri.me/problems/not_found"
          },
          "code": {
            "enum": [
              "bad_request",
              "invalid_fields",
              "invalid_id",
              "unauthenticated",
              "forbidden",
              "not_found",
              "conflict",
              "duplicate",
              "constraint_violation",
              "payload_too_large",
              "unsupported_media_type",
              "validation_failed",
              "check_violation",
              "rate_limited",
              "internal_error",
              "service_unavailable",
              "invalid_credentials",
              "invalid_token",
              "session_expired",
              "session_revoked",
              "wrong_current_password",
              "account_not_enabled",
              "account_pending_approval",
              "account_suspended",
              "credentials_not_managed",
              "real_accounts_protected",
              "weak_password",
              "identity_provider_failed",
              "user_already_active",
              "cannot_target_self",
              "last_active_admin",
              "password_reset_not_managed",
              "splits_unbalanced",
              "amount_breaks_splits",
              "transfer_same_account",
              "transfer_leg_locked",
              "account_not_owned",
              "category_not_owned",
              "account_has_transactions",
              "credit_fields_on_non_credit",
              "category_cycle",
              "category_too_deep",
              "category_has_children",
              "reassignment_required",
              "reassignment_target_inside",
              "template_requires_empty",
              "multi_payment_requires_concept",
              "multi_payment_requires_recurring",
              "multi_payment_excludes_auto_paid",
              "merge_into_itself",
              "merge_requires_concepts",
              "merge_source_has_children",
              "interpretation_needs_text",
              "concept_archived",
              "cost_center_cannot_classify",
              "no_files",
              "file_type_not_allowed",
              "file_content_mismatch",
              "file_too_large",
              "image_format_unsupported",
              "image_processing_unavailable",
              "receipt_file_missing",
              "database_unavailable"
            ],
            "type": "string",
            "description": "Stable, machine-readable: one per business rule (`problem-codes.ts`).",
            "example": "not_found"
          },
          "title": {
            "type": "string",
            "description": "Short summary of the kind of problem, in Spanish; the same for every occurrence."
          },
          "status": {
            "type": "number",
            "description": "The HTTP status, repeated."
          },
          "detail": {
            "type": "string",
            "description": "What happened this time, in Spanish: the client may show it as is."
          },
          "errors": {
            "description": "One per invalid field; present only when the input itself is at fault.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemFieldError"
            }
          }
        },
        "required": ["type", "code", "title", "status", "detail"]
      },
...
      "Split": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "categoryId": {
            "type": "number",
            "nullable": true
          },
          "amount": {
            "type": "string"
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["id", "categoryId", "amount", "note"]
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "type": {
            "enum": ["expense", "income", "transfer"],
            "type": "string"
          },
          "transferDirection": {
            "nullable": true,
            "enum": ["out", "in"],
            "type": "string"
          },
          "status": {
            "enum": ["cleared", "pending"],
            "type": "string"
          },
          "source": {
            "enum": ["web", "ios_manual", "ios_photo", "wallet", "sms"],
            "type": "string"
          },
          "id": {
            "type": "number"
          },
          "uuid": {
            "type": "string"
          },
          "accountId": {
            "type": "number",
            "nullable": true
          },
          "date": {
            "type": "string",
            "description": "`YYYY-MM-DD`."
          },
          "period": {
            "type": "string",
            "description": "The month the transaction BELONGS to, as its first day (`YYYY-MM-DD`)."
          },
          "amount": {
            "type": "string"
          },
          "currency": {
            "type": "string",
            "description": "ISO 4217 code of `amount`."
          },
          "categoryId": {
            "type": "number",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "merchant": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "transferGroupId": {
            "type": "string",
            "nullable": true
          },
          "externalRef": {
            "type": "string",
            "nullable": true,
            "description": "The client's own id, unique per user; a capture that repeats it gets this one back."
          },
          "rawText": {
            "type": "string",
            "nullable": true
          },
          "capturedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "needsReview": {
            "type": "boolean",
            "description": "Someone has to look at it: an unsure classification or a possible duplicate."
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "splits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Split"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "type",
          "transferDirection",
          "status",
          "source",
          "id",
          "uuid",
          "accountId",
          "date",
          "period",
          "amount",
          "currency",
          "categoryId",
          "description",
          "merchant",
          "notes",
          "transferGroupId",
          "externalRef",
          "rawText",
          "capturedAt",
          "needsReview",
          "tags",
          "splits",
          "createdAt"
        ]
      },
      "TransactionPageMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "1-based page number."
          },
          "perPage": {
            "type": "number",
            "description": "Items per page, 1–200."
          },
          "total": {
            "type": "number",
            "description": "Items across every page."
          },
          "sumExpense": {
            "type": "string",
            "description": "Sum of the expenses across every page that matches the filters."
          },
          "sumIncome": {
            "type": "string",
            "description": "Sum of the incomes across every page that matches the filters."
          }
        },
        "required": ["page", "perPage", "total", "sumExpense", "sumIncome"]
      },
...
      "CreateTransactionInput": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "amount": {
            "type": "string",
            "description": "Always positive: `type` gives the sign."
          },
          "accountId": {
            "type": "number",
            "description": "Optional: keeping accounts is a setting, not a requirement.",
            "minimum": 1
          },
          "categoryId": {
            "type": "number",
            "description": "Optional: a transaction may exist without a category."
          },
          "externalRef": {
            "type": "string",
            "description": "The client's own id, unique per user: a repeat here is a 409. To retry\nsafely, capture through `POST /transactions/capture`, which returns the\ntransaction already recorded.",
            "maxLength": 255
          },
          "period": {
            "type": "string",
            "description": "The month the transaction BELONGS to, as `YYYY-MM-DD` of its first day.\nAbsent, it is the month of `date`; it differs when a bill crosses months."
          },
          "rawText": {
            "type": "string",
            "nullable": true,
            "description": "The text it came from: the OCR of a receipt, a bank SMS.",
            "maxLength": 20000
          },
          "capturedAt": {
            "type": "string",
            "nullable": true,
            "description": "When it was captured, ISO 8601 with offset."
          },
          "needsReview": {
            "type": "boolean",
            "description": "Someone has to look at it: an unsure classification or a possible duplicate."
          },
          "type": {
            "enum": ["expense", "income", "transfer"],
            "type": "string"
          },
          "status": {
            "enum": ["cleared", "pending"],
            "type": "string"
          },
          "source": {
            "description": "Where it comes from: the web sends `web`, the phone app one of its own.",
            "enum": ["web", "ios_manual", "ios_photo", "wallet", "sms"],
            "type": "string"
          },
          "merchant": {
            "type": "string",
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "maxLength": 255
          },
          "notes": {
            "type": "string"
          },
          "tags": {
            "description": "Names, not ids: the tags are created on the fly.",
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "splits": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SplitInput"
            }
          }
        },
        "required": ["date", "amount"]
      },
...
      "Interpretation": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "nullable": true
          },
          "merchant": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "classification": {
            "$ref": "#/components/schemas/Classification"
          },
          "needsReview": {
            "type": "boolean"
          }
        },
        "required": ["amount", "date", "merchant", "description", "classification", "needsReview"]
      },
      "InterpretInput": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Free text: the OCR of a receipt, a bank SMS, what the user typed.",
            "maxLength": 20000
          },
          "merchant": {
            "type": "string",
            "maxLength": 255
          },
          "amount": {
            "type": "string",
            "description": "In pesos, up to two decimals; `.` or `,` as the separator.",
            "pattern": "^\\d+([.,]\\d{1,2})?$"
          },
          "date": {
            "type": "string",
            "description": "`YYYY-MM-DD`.",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "fileName": {
            "type": "string",
            "description": "Name of the file the text came from.",
            "maxLength": 255
          },
          "period": {
            "type": "string",
            "description": "The month it belongs to, `YYYY-MM`.",
            "pattern": "^\\d{4}-\\d{2}$"
          }
        }
      },
      "Capture": {
        "type": "object",
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/Transaction"
          },
          "classification": {
            "$ref": "#/components/schemas/Classification"
          },
          "summary": {
            "type": "string",
            "description": "One line for the notification, in Spanish."
          },
          "isDuplicate": {
            "type": "boolean",
            "description": "The `externalRef` was already captured: this is that transaction, unchanged."
          },
          "isMerged": {
            "type": "boolean",
            "description": "Merged into the other side of the same payment instead of creating a new one."
          }
        },
        "required": ["transaction", "classification", "summary", "isDuplicate", "isMerged"]
      },
...
    }
  }
}